Add ability to view tasks by completed date
Currently, there is no way of seeing when a task was completed,
only when it was created. In order to facilitate this, whenever an item
is checked, we set a completedDate. If the item is unchecked, we set the
completedDate to null. We also add a new flavor of the timeline view
which groups by completed date instead of by `_date`. This view
filters out any incomplete items (or complete items which do not have
the completedDate set).